+++ galite-core/src/main/kotlin/org/kopi/galite/visual/form/VColorField.kt
@@ -26,9 +26,7 @@ import kotlin.reflect.KClass
 
 import org.jetbrains.exposed.sql.Column
 import org.jetbrains.exposed.sql.ResultRow
-import org.jetbrains.exposed.sql.statements.api.ExposedBlob
 import org.kopi.galite.visual.VlibProperties
-import org.kopi.galite.visual.base.Utils
 import org.kopi.galite.visual.list.VColorColumn
 import org.kopi.galite.visual.list.VListColumn
 import org.kopi.galite.util.base.InconsistencyException
@@ -111,8 +109,8 @@ class VColorField(val bufferSize: Int) : VField(1,1) {
    * Warning:	This method will become inaccessible to users in next release
    */
   override fun setObject(r: Int, v: Any?) {
-    if (v is ByteArray) {
-      setColor(r, Color(reformat(v[0]), reformat(v[1]), reformat(v[2])))
+    if (v is org.kopi.galite.type.Color) {
+      setColor(r, Color(v.value))
     } else {
       setColor(r, v as Color?)
     }
@@ -124,11 +122,8 @@ class VColorField(val bufferSize: Int) : VField(1,1) {
    * @param    column       the column in the tuple
    */
   override fun retrieveQuery(result: ResultRow, column: Column<*>): Any? {
-    val bytes = when (val value = result[column]) {
-      is ExposedBlob -> {
-        value.bytes
-      }
-      is ByteArray -> {
+    val color = when (val value = result[column]) {
+      is org.kopi.galite.type.Color -> {
         value
       }
       else -> {
@@ -136,7 +131,7 @@ class VColorField(val bufferSize: Int) : VField(1,1) {
       }
     }
 
-    return Color(reformat(bytes[0]), reformat(bytes[1]), reformat(bytes[2]))
+    return color
   }
 
   /**
@@ -172,7 +167,7 @@ class VColorField(val bufferSize: Int) : VField(1,1) {
   /**
    * Returns the SQL representation of field value of given record.
    */
-  override fun getSqlImpl(r: Int): String? = if (value[r] == null) null else Utils.colorToRgbString(value[r])
+  override fun getSqlImpl(r: Int): org.kopi.galite.type.Color? = if (value[r] == null) null else org.kopi.galite.type.Color(value[r]!!.rgb)
 
   /**
    * Copies the value of a record to another
@@ -199,7 +194,7 @@ class VColorField(val bufferSize: Int) : VField(1,1) {
   /**
    * Warning:	This method will become inaccessible to users in next release
    */
-  override fun hasBinaryLargeObject(r: Int): Boolean = true
+  override fun hasBinaryLargeObject(r: Int): Boolean = false
 
   /**
    * Returns the SQL representation of field value of given record.
@@ -231,11 +226,6 @@ class VColorField(val bufferSize: Int) : VField(1,1) {
       true
     } ?: false
   }
-
-  /**
-   * Reformat a unsigned int from a byte
-   */
-  private fun reformat(b: Byte): Int = if (b < 0) b + 256 else b.toInt()
 
   /**
    * Get byteArray from Color