+++ galite-demo/galite-vaadin/src/main/kotlin/org/kopi/galite/demo/client/ClientForm.kt
@@ -36,6 +36,7 @@ import org.kopi.galite.visual.domain.STRING
 import org.kopi.galite.visual.dsl.common.Icon
 import org.kopi.galite.visual.dsl.common.Mode
 import org.kopi.galite.visual.dsl.form.Block
+import org.kopi.galite.visual.dsl.form.BlockOption
 import org.kopi.galite.visual.dsl.form.Border
 import org.kopi.galite.visual.dsl.form.DictionaryForm
 import org.kopi.galite.visual.dsl.form.FieldOption
@@ -140,6 +141,7 @@ class ClientForm : DictionaryForm(title = "Clients", locale = Locale.UK) {
      */
     fun save(b: VBlock) {
       clientsBlock.block.validate()
+      salesBlock.block.validate()
 
       if (!salesBlock.isFilled()) {
         salesBlock.currentRecord = 0
@@ -206,6 +208,7 @@ class ClientForm : DictionaryForm(title = "Clients", locale = Locale.UK) {
 
     init {
       border = Border.LINE
+      options(BlockOption.NODETAIL)
 
       command(item = showHideFilter) { showHideFilter() }
       command(item = report) { createReport { ClientR() } }